+++ packages/tools/helpers/src/actions.ts
@@ -8,7 +8,7 @@ import { generatePackageMetadata } from './generatePackageMetadata';
 
 const debug = buildDebug('verdaccio:tools:helpers:actions');
 
-export function publishVersion(app, pkgName, version, metadata: Partial<Manifest> = {}): any {
+export function publishVersion(app: any, pkgName, version, metadata: Partial<Manifest> = {}): any {
   debug('publishVersion %s : %s : %s', pkgName, version, JSON.stringify(metadata, null, 2));
   let pkgMetadata = { ...generatePackageMetadata(pkgName, version), ...metadata };
   // sync metadata readme to version of package
@@ -16,8 +16,8 @@ export function publishVersion(app, pkgName, version, metadata: Partial<Manifest
   debug('metadata %s', JSON.stringify(pkgMetadata, null, 2));
   return (
     supertest(app)
-      // @ts-ignore
       .put(`/${encodeURIComponent(pkgName)}`)
+      // @ts-ignore ignore content-type mismatch
       .set(HEADER_TYPE.CONTENT_TYPE, HEADERS.JSON)
       .send(JSON.stringify(pkgMetadata))
       .set('accept', HEADERS.GZIP)