+++ packages/web/test/api.packages.test.ts
@@ -4,6 +4,7 @@ import { afterEach, describe, expect, test, vi } from 'vitest';
 
 import { HEADERS, HEADER_TYPE, HTTP_STATUS } from '@verdaccio/core';
 import { setup } from '@verdaccio/logger';
+import { publishVersion } from '@verdaccio/test-helper';
 
 import { initializeServer } from './helper';
 
@@ -61,7 +62,8 @@ describe('test web server', () => {
       manifest: require('./partials/manifest/manifest.json'),
     }));
 
-    const api = supertest(await initializeServer('default-test.yaml'));
+    const app = await initializeServer('protected-package.yaml');
+    const api = supertest(app);
 
     // First login to get the token
     const loginRes = await api
@@ -77,6 +79,9 @@ describe('test web server', () => {
 
     expect(loginRes.body.token).toBeDefined();
 
+    // publish a package to be listed in the packages API
+    await publishVersion(app, 'foo', '1.0.0');
+
     // Then access the packages API with the token
     const response = await api
       .get('/-/verdaccio/data/packages')
@@ -85,6 +90,32 @@ describe('test web server', () => {
       .expect(HEADER_TYPE.CONTENT_TYPE, HEADERS.JSON_CHARSET)
       .expect(HTTP_STATUS.OK);
 
-    expect(response.body).toEqual([]);
+    expect(response.body.length).toEqual(1);
+  });
+
+  test('should not display package foo if user is logged out', async () => {
+    mockManifest.mockReturnValue(() => ({
+      staticPath: path.join(__dirname, 'static'),
+      manifestFiles: {
+        js: ['runtime.js', 'vendors.js', 'main.js'],
+      },
+      manifest: require('./partials/manifest/manifest.json'),
+    }));
+
+    const app = await initializeServer('protected-package.yaml');
+    const api = supertest(app);
+
+    // publish a protected package with 2 versions
+    await publishVersion(app, 'foo', '1.0.0');
+    await publishVersion(app, 'foo', '2.0.0');
+
+    // Then access the packages API with the token
+    const response = await api
+      .get('/-/verdaccio/data/packages')
+      .set('Accept', HEADERS.JSON_CHARSET)
+      .expect(HEADER_TYPE.CONTENT_TYPE, HEADERS.JSON_CHARSET)
+      .expect(HTTP_STATUS.OK);
+
+    expect(response.body.length).toEqual(0);
   });
 });